backend/requirements.txt
Flask==3.0.3
Flask-Cors==5.0.0
flask-smorest==0.44.0
python-dotenv==1.0.1
supabase==2.12.0
PyJWT==2.9.0
openai==1.55.0
pandas==2.2.2
pillow==10.3.0
Werkzeug==3.0.3
Flask-Limiter==3.5.0 